# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural drawing management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checklist-templates:
    get:
      operationId: listChecklistTemplates
      tags: [checklists]
      summary: List reusable checklist templates
      responses:
        "200":
          description: Checklist templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistTemplate"
    post:
      operationId: createChecklistTemplate
      tags: [checklists]
      summary: Create a reusable checklist template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTemplateInput"
      responses:
        "201":
          description: Created checklist template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTemplate"
        "409":
          description: A template with this name already exists
  /checklist-templates/{id}:
    patch:
      operationId: updateChecklistTemplate
      tags: [checklists]
      summary: Update a reusable checklist template
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTemplateInput"
      responses:
        "200":
          description: Updated checklist template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTemplate"
        "404":
          description: Checklist template not found
    delete:
      operationId: deleteChecklistTemplate
      tags: [checklists]
      summary: Delete a reusable checklist template
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted checklist template
        "404":
          description: Checklist template not found
  /project-checklists:
    get:
      operationId: listProjectChecklists
      tags: [checklists]
      summary: List checklists applied to projects
      parameters:
        - name: projectName
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Project checklists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectChecklist"
    post:
      operationId: applyChecklistTemplate
      tags: [checklists]
      summary: Apply a checklist template to a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectChecklistInput"
      responses:
        "201":
          description: Applied project checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectChecklist"
        "404":
          description: Checklist template not found
  /project-checklists/{id}:
    delete:
      operationId: deleteProjectChecklist
      tags: [checklists]
      summary: Remove a checklist from a project
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted project checklist
        "404":
          description: Project checklist not found
  /project-checklists/{id}/items/{itemId}:
    patch:
      operationId: toggleProjectChecklistItem
      tags: [checklists]
      summary: Complete or reopen a project checklist item
      parameters:
        - $ref: "#/components/parameters/Id"
        - name: itemId
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistItemUpdate"
      responses:
        "200":
          description: Updated project checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectChecklist"
        "404":
          description: Checklist item not found
  /chat/channels:
    get:
      operationId: listChatChannels
      tags: [chat]
      summary: List team chat channels
      responses:
        "200":
          description: Team chat channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatChannel"
    post:
      operationId: createChatChannel
      tags: [chat]
      summary: Create a team chat channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatChannelInput"
      responses:
        "201":
          description: Created chat channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatChannel"
        "409":
          description: A channel with this name already exists
  /chat/channels/{channelId}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List messages in a chat channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "404":
          description: Chat channel not found
    post:
      operationId: createChatMessage
      tags: [chat]
      summary: Post a message to a chat channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Created chat message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "404":
          description: Chat channel not found
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
...
  schemas:
...
    ChecklistTemplateItem:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        position:
          type: number
      required: [id, title, position]
    ChecklistTemplate:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        createdBy:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistTemplateItem"
      required: [id, name, description, createdBy, createdAt, updatedAt, items]
    ChecklistTemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        items:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
      required: [name, items]
    ProjectChecklistItem:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        position:
          type: number
        completed:
          type: boolean
        completedBy:
          type: ["number", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
      required: [id, title, position, completed, completedBy, completedAt]
    ProjectChecklist:
      type: object
      properties:
        id:
          type: number
        projectName:
          type: string
        templateId:
          type: number
        name:
          type: string
        createdBy:
          type: number
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectChecklistItem"
      required: [id, projectName, templateId, name, createdBy, createdAt, items]
    ProjectChecklistInput:
      type: object
      properties:
        projectName:
          type: string
          minLength: 1
        templateId:
          type: number
          minimum: 1
      required: [projectName, templateId]
    ChecklistItemUpdate:
      type: object
      properties:
        completed:
          type: boolean
      required: [completed]
...
    ChatChannel:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        createdBy:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, createdBy, createdAt]
    ChatChannelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required: [name]
    ChatMessage:
      type: object
      properties:
        id:
          type: number
        channelId:
          type: number
        authorId:
          type: number
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, channelId, authorId, authorName, content, createdAt]
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]
...